moved create form into its own component, changed obj to a state for easier understanding
