+++ gracefabrics/src/pages/productManagement.js
@@ -1,6 +1,7 @@
 import { React, useEffect, useState } from 'react';
 import ManageProduct from '../components/ManageProduct';
-import { Grid, Button, makeStyles, TextField } from '@material-ui/core';
+import CreateProductForm from '../components/CreateProductForm';
+import { Grid, Button} from '@material-ui/core';
 import API from '../utils/API';
 
 const ProductManagement = () => {
@@ -21,17 +22,6 @@ const ProductManagement = () => {
         productsList.renderProducts()
     }, []);
 
-
-    function createRender(event) {
-        API.createProduct(event).then(() => {
-            document.querySelectorAll("input").forEach(function (input) {
-                input.value = ""
-            })
-            alert("New Product Created --Scroll to the bottom to see :)")
-            productsList.renderProducts()
-        })
-    }
-
     const productsToMap = productsList.products.map(product => {
         return (
             <Grid item xs={12} sm={12} md={12} key={product.id}>
@@ -43,30 +33,6 @@ const ProductManagement = () => {
             </Grid>
         )
     })
-
-    const useStyles = makeStyles((theme) => ({
-        root: {
-            '& > *': {
-                margin: theme.spacing(1),
-                width: '25ch',
-            },
-        },
-    }));
-
-    const classes = useStyles();
-
-    let Products = {
-        name: "",
-        type: "",
-        price: "",
-        description: "",
-        img1: ""
-    }
-    function updateProductsObj(e) {
-        Products[e.target.id] = e.target.value
-        document.getElementById("CreateButton").value = JSON.stringify(Products)
-    }
-
     return (
         <div >
             <div className="blurb">
@@ -75,24 +41,7 @@ const ProductManagement = () => {
                 </h1>
                 <Button onClick={()=>{window.location.replace("/")}}>Log Out</Button>
             </div>
-            <form className={classes.root} noValidate autoComplete="off">
-                <table>
-                    <tbody>
-                        <tr>
-                            <td><TextField id="name" label="Name" variant="outlined" onChange={updateProductsObj} /></td>
-                            <td><TextField id="type" label="Category" variant="outlined" onChange={updateProductsObj} /></td>
-                        </tr>
-                        <tr>
-                            <td><TextField id="price" label="Price" variant="outlined" onChange={updateProductsObj} /></td>
-                            <td><TextField id="description" label="Description" variant="outlined" onChange={updateProductsObj} /></td>
-                        </tr>
-                        <tr>
-                            <td><TextField id="img1" label="URL to Image" variant="outlined" onChange={updateProductsObj} /></td>
-                            <td><Button id="CreateButton" onClick={createRender} value={JSON.stringify(Products)}>Create New Item</Button></td>
-                        </tr>
-                    </tbody>
-                </table>
-            </form>
+            <CreateProductForm renderProducts = {productsList.renderProducts}/>
             <hr />
             <Grid container>
                 {productsToMap}
